+++ lib/sails/config.rb
@@ -13,7 +13,11 @@ module Sails
       config.i18n = I18n
       config.i18n.load_path += Dir[Sails.root.join('config', 'locales', '*.{rb,yml}').to_s]
       config.i18n.default_locale = :en
-      config.thrift_processor = nil
+      
+      config.port = 4000
+      config.thread_port = 4001
+      config.processor = nil
+      config.protocol = :binary
     end
   end
 end