+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `sails s`, `sails c` commands.
 - Refactor service layer, use class to instead module.
+- Support to custom Thrift protocol with `config.protocol = :binary`.
 - Fix ThriftServer::OperationFailed not found error.
 
 ## 0.1.0 / 2014-12-9